refactor(resume): move command menu and project card to components folder
