+++ app/components/command-menu.tsx
@@ -11,7 +11,7 @@ import {
   CommandList,
   CommandSeparator,
 } from '@/components/ui/command'
-import { Button } from './ui/button'
+import { Button } from '../../components/ui/button'
 import { CommandIcon } from 'lucide-react'
 
 interface Props {
@@ -36,11 +36,11 @@ export const CommandMenu = ({ links }: Props) => {
   return (
     <>
       <p className="fixed bottom-0 left-0 right-0 hidden border-t border-t-muted bg-white dark:bg-gray-900 p-1 text-center text-sm text-muted-foreground print:hidden xl:block">
-        Press{' '}
+        Pressione{' '}
         <kbd className="pointer-events-none inline-flex h-5 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium text-muted-foreground opacity-100">
           <span className="text-xs">⌘</span>J
         </kbd>{' '}
-        to open the command menu
+        para abrir o menu de comandos
       </p>
       <Button
         onClick={() => setOpen((open) => !open)}
@@ -51,14 +51,14 @@ export const CommandMenu = ({ links }: Props) => {
         <CommandIcon className="my-6 size-6" />
       </Button>
       <CommandDialog open={open} onOpenChange={setOpen}>
-        <CommandInput placeholder="Type a command or search..." />
+        <CommandInput placeholder="Selecione ou pesquise um comando..." />
         <CommandList>
-          <CommandEmpty>No results found.</CommandEmpty>
+          <CommandEmpty>Nenhum resultado encontrado.</CommandEmpty>
           <CommandGroup heading="Actions">
             <CommandItem
               onSelect={() => {
                 setOpen(false)
-                window.print()
+                setTimeout(() => window.print(), 0)
               }}
             >
               <span>Print</span>
